+++ src/modules/ocpOverviewWidget/ocpOverviewChart.tsx
@@ -15,7 +15,7 @@ import { createMapStateToProps, FetchStatus } from 'store/common';
 import { reportActions, reportSelectors } from 'store/reports';
 import { unitLookupKey } from 'utils/formatValue';
 import { formatValue } from 'utils/formatValue';
-import { skeletonWidth } from 'utils/skeletonSize';
+import { skeletonWidth } from 'utils/skeleton';
 
 import { chartStyles, styles } from './ocpOverviewChart.styles';
